requirements.txt
Django
python-environ
django-crispy-forms